+++ PCL.Neo.Core/Models/Minecraft/Game/Versions.cs
@@ -1,0 +1,6 @@
+namespace PCL.Neo.Core.Models.Minecraft.Game
+{
+    internal class Versions
+    {
+    }
+}
