+++ PCL.Neo.Core/Models/Minecraft/Java/IJavaManager.cs
@@ -1,0 +1,9 @@
+namespace PCL.Neo.Core.Models.Minecraft.Java;
+
+public interface IJavaManager
+{
+    List<JavaRuntime> JavaList { get; }
+    Task JavaListInit();
+    Task ManualAdd(string javaDir);
+    Task Refresh();
+}
