--- PCL.Neo.Tests/Models/Minecraft/JavaTest.cs
@@ -1,15 +1,0 @@
-using PCL.Neo.Models.Minecraft.Java;
-using PCL.Neo.Utils;
-
-namespace PCL.Neo.Tests.Models.Minecraft
-{
-    public class JavaTest
-    {
-        [Test]
-        public async Task Test()
-        {
-            JavaManager javaInstance = new();
-            await javaInstance.JavaListInit();
-        }
-    }
-}
