--- PCL.Neo/Models/Minecraft/ReleaseTypeEnum.cs
@@ -1,13 +1,0 @@
-using System.Text.Json.Serialization;
-
-namespace PCL.Neo.Models.Minecraft;
-
-[JsonConverter(typeof(JsonStringEnumConverter))]
-public enum ReleaseTypeEnum
-{
-    Unknown,
-    [JsonStringEnumMemberName("release")] Release,
-    [JsonStringEnumMemberName("snapshot")] Snapshot,
-    [JsonStringEnumMemberName("old_alpha")] OldAlpha,
-    [JsonStringEnumMemberName("old_beta")] OldBeta
-}
