+++ PCL.Neo/PCL.Neo.csproj
@@ -68,6 +68,9 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Folder Include="Models\Minecraft\Game\Data\" />
+    <Folder Include="Models\Minecraft\Java\" />
+    <Folder Include="Models\Minecraft\Mod\" />
     <Folder Include="Views\MainPage\" />
     <Folder Include="Views\SubPage\" />
   </ItemGroup>
